+++ src/editor-mathfield/styling.ts
@@ -38,13 +38,16 @@ export function applyStyle(
     // This global style will be used the next time an atom is inserted
     mathfield.style = { ...mathfield.style, ...style };
   } else {
-    mathfield.model.deferNotifications({ content: true }, () => {
-      // Change the style of the selection
-      model.selection.ranges.forEach((range) =>
-        applyStyleToModel(model, range, style, { operation: 'toggle' })
-      );
-      mathfield.snapshot();
-    });
+    mathfield.model.deferNotifications(
+      { content: true, type: 'insertText' },
+      () => {
+        // Change the style of the selection
+        model.selection.ranges.forEach((range) =>
+          applyStyleToModel(model, range, style, { operation: 'toggle' })
+        );
+        mathfield.snapshot();
+      }
+    );
   }
 
   return true;