+++ src/editor-model/commands-delete.ts
@@ -2,38 +2,49 @@ import { register } from '../editor/commands';
 import type { ModelPrivate } from './model-private';
 import { deleteForward, deleteBackward, deleteRange } from './delete';
 import { wordBoundaryOffset } from './commands';
+import { contentWillChange } from './listeners';
 
 register(
   {
-    deleteAll: (model: ModelPrivate): boolean => {
-      return deleteRange(model, [0, -1]);
-    },
+    deleteAll: (model: ModelPrivate): boolean =>
+      contentWillChange(model, { inputType: 'deleteContent' }) &&
+      deleteRange(model, [0, -1], 'deleteContent'),
     deleteForward: (model: ModelPrivate): boolean => deleteForward(model),
     deleteBackward: (model: ModelPrivate): boolean => deleteBackward(model),
     deleteNextWord: (model: ModelPrivate): boolean =>
-      deleteRange(model, [
-        model.anchor,
-        wordBoundaryOffset(model, model.position, 'forward'),
-      ]),
+      contentWillChange(model, { inputType: 'deleteWordForward' }) &&
+      deleteRange(
+        model,
+        [model.anchor, wordBoundaryOffset(model, model.position, 'forward')],
+        'deleteWordForward'
+      ),
     deletePreviousWord: (model: ModelPrivate): boolean =>
-      deleteRange(model, [
-        model.anchor,
-        wordBoundaryOffset(model, model.position, 'backward'),
-      ]),
+      contentWillChange(model, { inputType: 'deleteWordBackward' }) &&
+      deleteRange(
+        model,
+        [model.anchor, wordBoundaryOffset(model, model.position, 'backward')],
+        'deleteWordBackward'
+      ),
     deleteToGroupStart: (model: ModelPrivate): boolean =>
-      deleteRange(model, [
-        model.anchor,
-        model.offsetOf(model.at(model.position).firstSibling),
-      ]),
+      contentWillChange(model, { inputType: 'deleteSoftLineBackward' }) &&
+      deleteRange(
+        model,
+        [model.anchor, model.offsetOf(model.at(model.position).firstSibling)],
+        'deleteSoftLineBackward'
+      ),
     deleteToGroupEnd: (model: ModelPrivate): boolean =>
-      deleteRange(model, [
-        model.anchor,
-        model.offsetOf(model.at(model.position).lastSibling),
-      ]),
+      contentWillChange(model, { inputType: 'deleteSoftLineForward' }) &&
+      deleteRange(
+        model,
+        [model.anchor, model.offsetOf(model.at(model.position).lastSibling)],
+        'deleteSoftLineForward'
+      ),
     deleteToMathFieldStart: (model: ModelPrivate): boolean =>
-      deleteRange(model, [model.anchor, 0]),
+      contentWillChange(model, { inputType: 'deleteHardLineBackward' }) &&
+      deleteRange(model, [model.anchor, 0], 'deleteHardLineBackward'),
     deleteToMathFieldEnd: (model: ModelPrivate): boolean =>
-      deleteRange(model, [model.anchor, -1]),
+      contentWillChange(model, { inputType: 'deleteHardLineForward' }) &&
+      deleteRange(model, [model.anchor, -1], 'deleteHardLineForward'),
   },
   { target: 'model', category: 'delete' }
 );